requirements.txt
albumentations==1.3.0
appdirs==1.4.4
asttokens==2.2.1
backcall==0.2.0
certifi==2022.12.7
cfgv==3.3.1
charset-normalizer==3.1.0
click==8.1.3
cmake==3.26.0
comm==0.1.2
contourpy==1.0.7
cycler==0.11.0
debugpy==1.6.6
decorator==5.1.1
distlib==0.3.6
docker-pycreds==0.4.0
executing==1.2.0
filelock==3.9.1
fonttools==4.39.0
gitdb==4.0.10
GitPython==3.1.31
huggingface-hub==0.13.2
identify==2.5.20
idna==3.4
imageio==2.26.0
importlib-metadata==6.0.0
importlib-resources==5.12.0
ipykernel==6.21.3
ipython==8.11.0
ipywidgets==8.0.4
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jupyter_client==8.0.3
jupyter_core==5.2.0
jupyterlab-widgets==3.0.5
kiwisolver==1.4.4
kornia==0.6.10
lazy_loader==0.1
lit==15.0.7
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mpmath==1.3.0
nest-asyncio==1.5.6
networkx==3.0
nodeenv==1.7.0
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
opencv-python-headless==4.7.0.72
packaging==23.0
pandas==1.5.3
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.1.1
pre-commit==3.1.1
prompt-toolkit==3.0.38
protobuf==4.22.1
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.14.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==25.0.1
qudida==0.0.4
regex==2022.10.31
requests==2.28.2
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.9.1
seaborn==0.12.2
sentry-sdk==1.17.0
setproctitle==1.3.2
six==1.16.0
smmap==5.0.0
stack-data==0.6.2
sympy==1.11.1
termcolor==2.2.0
threadpoolctl==3.1.0
tifffile==2023.2.28
tokenizers==0.13.2
torch==2.0.0
torchaudio==2.0.1
torchgeometry==0.1.2
torchmetrics==0.11.4
torchvision==0.15.1
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.27.1
triton==2.0.0
typing_extensions==4.5.0
urllib3==1.26.15
virtualenv==20.21.0
wandb==0.14.0
wcwidth==0.2.6
widgetsnbextension==4.0.5
zipp==3.15.0
